randgen: extract table logic into stand alone file
Previously, the logic for generating schemas was mixed up with the logic
for generating tables. This moves table specific logic into its own
file.

Release note: none
